+++ src/inputplugins/packer.cpp
@@ -1602,7 +1602,7 @@ CIPPacker::getRandVelVec()
     normal_sampler(Sim->ranGenerator, normdist);
   
   Vector  tmpVec;
-  for (int iDim = 0; iDim < NDIM; iDim++)
+  for (size_t iDim = 0; iDim < NDIM; iDim++)
     tmpVec[iDim] = normal_sampler();
   
   return tmpVec;
