+++ src/outputplugins/correlations/thermalCondSpeciesSpeciesE.cpp
@@ -148,7 +148,7 @@ COPThermalConductivitySpeciesSpeciesE::output(xmlw::XmlStream &XML)
 	    XML << (1+i) * dt / Sim->Dynamics.units().unitTime()
 		<< "\t ";
 	    
-	    for (int j=0;j<NDIM;j++)
+	    for (size_t j=0;j<NDIM;j++)
 	      XML << accG2[id1 + Nsp * id2][i][j] * factor 
 		  << "\t ";
